refactor(fwa): rename Points command file to Fwa and update imports
