+++ src/Commands.ts
@@ -8,7 +8,7 @@ import { Compo } from "./commands/Compo";
 import { Post } from "./commands/Post";
 import { CommandRole } from "./commands/CommandRole";
 import { CC } from "./commands/CC";
-import { Fwa } from "./commands/Points";
+import { Fwa } from "./commands/Fwa";
 import { Opponent } from "./commands/Opponent";
 import { KickList } from "./commands/KickList";
 import { Recruitment } from "./commands/Recruitment";
